requirements.txt
# core data processing
pandas>=2.0.0
numpy>=1.24.0

# Google Cloud and BigQuery
google-cloud-bigquery>=3.11.0
google-auth>=2.17.0
google-auth-oauthlib>=0.1.0

# configuration
dotenv>=0.9.9
python-dotenv>=1.1.1
pyyaml>=6.0.2

# utilities
structlog>=25.4.0